Remove unused edge runtime and reorganize components
Deleted the runtime constraints from multiple files as they are no longer needed. Reorganized component imports and moved several files to improve project structure and maintainability.
