+++ apps/web/src/app/api/auth/[...nextauth]/route.ts
@@ -1,4 +1,3 @@
 import { handlers } from '@/lib/auth';
 
-export const runtime = "edge";
 export const { GET, POST } = handlers;
