+++ apps/web/src/app/api/auth/logout/route.ts
@@ -1,6 +1,4 @@
 import { NextResponse } from "next/server";
-
-export const runtime = "edge";
 export async function GET(request: Request) {
   const requestUrl = new URL(request.url);
   const origin = requestUrl.origin;
